+++ processor.py
@@ -656,10 +656,8 @@ def handle_export_workflows(timeout=120):
             return
         
         # If succeed
-        sobjects_describe = api.result
-        for sobject in sobjects_describe:
-            util.parse_workflow_metadata(toolingapi_settings, sobject)
-
+        sobjects = api.result.keys()
+        util.parse_workflow_metadata(toolingapi_settings, sobjects)
         print (message.SEPRATE.format("Outputdir: " + outputdir))
 
     toolingapi_settings = context.get_toolingapi_settings()
@@ -668,27 +666,6 @@ def handle_export_workflows(timeout=120):
     thread = threading.Thread(target=api.describe_global, args=())
     thread.start()
     ThreadProgress(api, thread, "Export All Workflows", "Outputdir: " + outputdir)
-    handle_thread(thread, 10)
-
-def handle_export_field_dependencies(timeout=120):
-    def handle_thread(thread, timeout):
-        if thread.is_alive():
-            sublime.set_timeout(lambda: handle_thread(thread, timeout), timeout)
-            return
-        
-        # If succeed
-        sobjects_describe = api.result
-        for sobject in sobjects_describe:
-            util.parse_field_dependencies(toolingapi_settings, sobject)
-
-        print (message.SEPRATE.format("Outputdir: " + outputdir))
-
-    toolingapi_settings = context.get_toolingapi_settings()
-    outputdir = toolingapi_settings["workspace"] + "/fieldDependencies/"
-    api = SalesforceApi(toolingapi_settings)
-    thread = threading.Thread(target=api.describe_global, args=())
-    thread.start()
-    ThreadProgress(api, thread, "Export All Sobject Field Dependencies", "Outputdir: " + outputdir)
     handle_thread(thread, 10)
 
 def handle_export_validation_rules(timeout=120):